More refactoring :(
